update-index: add untracked cache notifications
Attempting to flip the untracked-cache feature on for a random index
file with

    cd /random/unrelated/place
    git --git-dir=/somewhere/else/.git update-index --untracked-cache

would not work as you might expect. Because flipping the feature on
in the index also records the location of the corresponding working
tree (/random/unrelated/place in the above example), when the index
is subsequently used to keep track of files in the working tree in
/somewhere/else, the feature is disabled.

With this patch "git update-index --untracked-cache" tells the user in
which directory tests are performed. This makes it easy to spot any
problem.

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>