this sort of config struct makes sense
If you set the flag to enabled we'll just infer the conf_key by
default, need something to pass us the command name somehow. We should
have it, see how --help finds the manpage.

The conf_callback would then be expected to munge the config value in
the same way as the command-line value.

Most of the time the callback is the exact same as the `callback`,
e.g. parse_opt_rebase.