grep: add support for PCRE v2
Add support for v2 of the PCRE API. This is a new major version of
PCRE that came out in early 2015[1].

The regular expression syntax is the same, but while the API is
similar, pretty much every function is either renamed or takes
different arguments. Thus using it via entirely new functions makes
sense, as opposed to trying to e.g. have one compile_pcre_pattern()
that would call either PCRE v1 or v2 functions.

Git can now be compiled with either USE_LIBPCRE1=[YesPlease|] or
USE_LIBPCRE2=[YesPlease|], with USE_LIBPCRE=[YesPlease|] currently
being a synonym for the former. Providing both is an error.

With earlier patches to enable JIT for PCRE v1 the performance of the
release versions of both libraries is almost exactly the same
performance, with PCRE v2 being around 1% slower.

However after I reported this to the pcre-dev mailing list[2] I got a
lot of help with the API use from Zoltán Herczeg, he subsequently
optimized some of the JIT functionality in v2 of the library.

Running the p7820-grep-engines.sh performance test against the latest
Subversion trunk of both, with both them and git compiled as -O3, and
the test run against linux.git, gives the following results:

    $ GIT_PERF_REPEAT_COUNT=30 GIT_PERF_LARGE_REPO=~/g/linux GIT_PERF_MAKE_COMMAND='grep -q LIBPCRE2 Makefile && make -j8 USE_LIBPCRE2=YesPlease CC=~/perl5/installed/bin/gcc NO_R_TO_GCC_LINKER=YesPlease CFLAGS=-O3 LIBPCREDIR=/home/avar/g/pcre2/inst LDFLAGS=-Wl,-rpath,/home/avar/g/pcre2/inst/lib || make -j8 USE_LIBPCRE=YesPlease CC=~/perl5/installed/bin/gcc NO_R_TO_GCC_LINKER=YesPlease CFLAGS=-O3 LIBPCREDIR=/home/avar/g/pcre/inst LDFLAGS=-Wl,-rpath,/home/avar/g/pcre/inst/lib' ./run HEAD~2 HEAD~ HEAD p7820-grep-engines.sh
    [...]
    Test                                           HEAD~2            HEAD~                    HEAD
    ----------------------------------------------------------------------------------------------------------------
    7820.1: basic grep how.to                      0.28(1.25+0.42)   0.28(1.11+0.55) +0.0%    0.28(1.24+0.46) +0.0%
    7820.2: extended grep how.to                   0.27(1.26+0.40)   0.28(1.10+0.56) +3.7%    0.29(1.18+0.49) +7.4%
    7820.3: perl grep how.to                       0.26(1.22+0.38)   0.19(0.36+0.60) -26.9%   0.19(0.30+0.62) -26.9%
    7820.5: basic grep ^how to                     0.27(1.21+0.42)   0.27(1.25+0.38) +0.0%    0.28(1.28+0.36) +3.7%
    7820.6: extended grep ^how to                  0.27(1.20+0.45)   0.28(1.24+0.42) +3.7%    0.29(1.26+0.44) +7.4%
    7820.7: perl grep ^how to                      0.48(2.75+0.40)   0.23(0.64+0.54) -52.1%   0.18(0.29+0.62) -62.5%
    7820.9: basic grep [how] to                    0.41(2.20+0.42)   0.42(2.28+0.31) +2.4%    0.42(2.26+0.36) +2.4%
    7820.10: extended grep [how] to                0.42(2.22+0.40)   0.41(2.16+0.44) -2.4%    0.42(2.16+0.47) +0.0%
    7820.11: perl grep [how] to                    0.56(2.90+0.48)   0.29(0.99+0.43) -48.2%   0.23(0.40+0.59) -58.9%
    7820.13: basic grep \(e.t[^ ]*\|v.ry\) rare    0.55(3.28+0.42)   0.54(3.30+0.40) -1.8%    0.56(3.30+0.41) +1.8%
    7820.14: extended grep (e.t[^ ]*|v.ry) rare    0.56(3.31+0.38)   0.63(3.44+0.59) +12.5%   0.56(3.35+0.37) +0.0%
    7820.15: perl grep (e.t[^ ]*|v.ry) rare        0.88(5.72+0.43)   0.27(1.13+0.46) -69.3%   0.20(0.48+0.58) -77.3%
    7820.17: basic grep m\(ú\|u\)ult.b\(æ\|y\)te   0.28(1.32+0.43)   0.28(1.33+0.41) +0.0%    0.30(1.31+0.46) +7.1%
    7820.18: extended grep m(ú|u)ult.b(æ|y)te      0.28(1.27+0.47)   0.28(1.30+0.42) +0.0%    0.30(1.27+0.48) +7.1%
    7820.19: perl grep m(ú|u)ult.b(æ|y)te          0.31(1.54+0.49)   0.22(0.70+0.52) -29.0%   0.18(0.32+0.59) -41.9%

Here HEAD~2 is git with PCRE v1 without JIT, HEAD~ is PCRE v1 with
JIT, and HEAD is PCRE v2 (also with JIT). See previous commits of mine
mentioning p7820-grep-engines.sh for more details on the test setup.

For ease of readability, a different run just of HEAD~ (PCRE v1 with
JIT) v.s. PCRE v2:

    Test                                           HEAD~             HEAD
    ---------------------------------------------------------------------------------------
    7820.1: basic grep how.to                      0.28(1.29+0.38)   0.28(1.22+0.44) +0.0%
    7820.2: extended grep how.to                   0.28(1.16+0.50)   0.28(1.22+0.44) +0.0%
    7820.3: perl grep how.to                       0.22(0.35+0.54)   0.19(0.34+0.58) -13.6%
    7820.5: basic grep ^how to                     0.27(1.20+0.43)   0.27(1.22+0.40) +0.0%
    7820.6: extended grep ^how to                  0.27(1.24+0.39)   0.27(1.22+0.42) +0.0%
    7820.7: perl grep ^how to                      0.21(0.69+0.48)   0.18(0.29+0.61) -14.3%
    7820.9: basic grep [how] to                    0.42(2.27+0.33)   0.41(2.19+0.43) -2.4%
    7820.10: extended grep [how] to                0.41(2.18+0.43)   0.42(2.23+0.38) +2.4%
    7820.11: perl grep [how] to                    0.26(0.92+0.50)   0.21(0.40+0.59) -19.2%
    7820.13: basic grep \(e.t[^ ]*\|v.ry\) rare    0.54(3.29+0.39)   0.54(3.32+0.39) +0.0%
    7820.14: extended grep (e.t[^ ]*|v.ry) rare    0.54(3.31+0.39)   0.54(3.20+0.50) +0.0%
    7820.15: perl grep (e.t[^ ]*|v.ry) rare        0.26(1.07+0.51)   0.20(0.52+0.54) -23.1%
    7820.17: basic grep m\(ú\|u\)ult.b\(æ\|y\)te   0.28(1.41+0.31)   0.28(1.27+0.47) +0.0%
    7820.18: extended grep m(ú|u)ult.b(æ|y)te      0.28(1.32+0.43)   0.28(1.31+0.43) +0.0%
    7820.19: perl grep m(ú|u)ult.b(æ|y)te          0.22(0.73+0.49)   0.18(0.32+0.59) -18.2%

I.e. the two are either neck-to-neck, but PCRE v2 usually pulls ahead,
when it does it's up to 20% faster.

A brief note on thread safety: As noted in pcre2api(3) & pcre2jit(3)
the compiled pattern can be shared between threads, but not some of
the JIT context, however the grep threading support does all pattern &
JIT compilation in separate threads, so this code doesn't need to
concern itself with thread safety.

See commit 63e7e9d8b6 ("git-grep: Learn PCRE", 2011-05-09) for the
initial addition of PCRE v1. This change follows some of the same
patterns it did (and which were discussed on list at the time),
e.g. mocking up types with typedef instead of ifdef-ing them out when
USE_LIBPCRE2 isn't defined. This adds some trivial memory use to the
program, but makes the code look nicer.

1. https://lists.exim.org/lurker/message/20150105.162835.0666407a.en.html
2. https://lists.exim.org/lurker/thread/20170419.172322.833ee099.en.html

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>